moved html assembly to build.ts
