+++ common.ts
@@ -23,28 +23,3 @@ export const get_html = async (path: string): Promise<string> => {
     
     return content_as_html;
 };
-
-export const assemble_html_page = (content: string, stylesheet: string) => {
-
-    if (!stylesheet.endsWith(".css")) throw "stylesheet name has to end with .css"
-
-    const template = `
-        <!DOCTYPE html>
-        <html lang="en">
-        <head>
-            <meta charset="UTF-8">
-            <meta name="viewport" content="width=device-width, initial-scale=1.0">
-            <meta http-equiv="X-UA-Compatible" content="ie=edge">
-            <title>Blog</title>
-            <!--Default stylesheet:-->
-            <link rel="stylesheet" type="text/css" href="MB_STYLESHEET">
-        </head>
-        <body>
-            MB_CONTENT
-        </body>
-        </html>
-    `
-    return template
-        .replace("MB_CONTENT", content)
-        .replace("MB_STYLESHEET", stylesheet); 
-}
