+++ deps.ts
@@ -5,5 +5,4 @@ export { decode } from "https://deno.land/std/strings/decode.ts";
 export { encode } from "https://deno.land/std/strings/encode.ts";
 export { green, red, yellow, bold } from "https://deno.land/std/fmt/mod.ts";
 
-export { suite } from "https://gist.githubusercontent.com/olaven/99bfe1628cef08060b438dd6634bb376/raw/0c7e8c5d2ec51b221b9230eaba9cff7e2612b9df/suite.ts"
 export { assertEquals, assert, assertThrows } from "https://deno.land/std/testing/asserts.ts";
